requirements.txt
astroid==2.0.4
autopep8==1.4.3
certifi==2018.10.15
chardet==3.0.4
Django==2.1.3
django-filter==2.0.0
django-oauth-toolkit==1.2.0
djangorestframework==3.9.0
docutils==0.14
idna==2.7
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
oauthlib==2.1.0
Pillow==5.3.0
pycodestyle==2.4.0
pylint==2.1.1
pytz==2018.7
qrcode==6.0
requests==2.20.1
six==1.11.0
typed-ast==1.1.0
urllib3==1.24.1
wrapt==1.10.11
